chiminey-beat/requirements.txt
Django==1.6.10
#SPARQLWrapper==1.6.4 #apt: 1.4.1
South==1.0.2 #apt: 0.7.5
Sphinx
Wand
amqp==1.4.6 #apt: 1.3.3
argparse==1.2.1 #apt: libpython2.7-stdlib
billiard
bleach==1.4.1 #apt: 1.2.2
celery==3.1.17 #apt: 3.1.6
compare
django-bootstrap-form==2.0.6 #apt: 3.1.0
django-celery==3.1.16 #apt: 2.5.5
django-extensions==1.1.1 #apt: 1.2.5
django-form-utils==0.2.0
#django-haystack==1.2.7
django-jasmine==0.4.1
django-mustachejs==0.6.0
django-nose
django-registration==1.0 #apt: 1.0
-e git+https://github.com/grischa/django-tastypie.git@testing-mytardis-deployment#egg=django-tastypie
django-user-agents==0.2.2
ecdsa==0.11 #apt: 0.10
#elementtree
gunicorn
gevent
html2text
isodate==0.5.1 #apt: 0.4.6
pyjwt==1.0.1
kombu==3.0.24 #apt: 3.0.7
lxml==3.2.1 #apt: 3.3.3
mimeparse==0.1.3 #apt: 0.1.4
nose
nosexcover
paramiko==1.15.2 #apt: 1.10.1
pwgen==0.4
pylint
pyoai==2.4.4
pyparsing==1.5.7 #apt: 2.0.1
pysolr>=2.1.0,<3 #apt: 2.0.15
python-dateutil==2.4.0 #apt: 1.5
# python-magic==0.4.6 #apt: 1:5.14
-e git+https://github.com/grischa/python-magic.git#egg=python-magic
pytz==2013b #apt: 2012c
rdfextras==0.4
rdflib==4.0.1 #apt: 2.4.2
six==1.9.0 #apt: 1.5.2
ua-parser==0.3.3
user-agents==0.1.1
wsgiref==0.1.2

# Chiminey extras
mock==1.0.1
flexmock
django-inmemorystorage
HTTPretty
docutils
django-widget-tweaks
django-storages==1.1.8
boto
#django-celery-with-redis==3.0
#redis==2.8.0
#hiredis==0.1.1
django-celery-with-redis
redis
hiredis



